+++ examplepatches/MiguelNinja19/miguel-morphe-patches/patches/src/main/kotlin/air/com/midjiwan/polytopia/patches/shared/Constants.kt
@@ -1,8 +1,8 @@
 package air.com.midjiwan.polytopia.patches.shared
 
-import app.morphe.patcher.patch.ApkFileType
-import app.morphe.patcher.patch.AppTarget
 import app.morphe.patcher.patch.Compatibility
+import app.morphe.patcher.patch.AppTarget
+import app.morphe.patcher.patch.ApkFileType
 
 val POLYTOPIA = Compatibility(
     name = "The Battle of Polytopia",